Move testGHOSTDagSorter to testutils
